requirements.txt
aicsimageio==4.14.0
aicspylibczi==3.2.0
aiobotocore==2.5.4
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
bokeh==3.6.0
botocore==1.31.17
cellpose==3.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
czifile==2019.7.2
dask==2024.10.0
dask-expr==1.1.16
dask-image==2024.5.3
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distributed==2024.10.0
elementpath==4.6.0
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
fastjsonschema==2.20.0
fastremap==1.15.0
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.6.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imagecodecs==2024.9.22
imageio==2.36.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
locket==1.0.0
lxml==4.9.4
lz4==4.3.3
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numcodecs==0.13.1
numpy==1.26.4
ome-types==0.5.2
ome-zarr==0.9.0
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.0.0
PIMS==0.7
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-compat==0.1.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
resource-backed-dask-array==0.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roifile==2024.9.15
rpds-py==0.20.1
s3fs==2023.6.0
scikit-image==0.24.0
scipy==1.14.1
Send2Trash==1.8.3
six==1.16.0
slicerator==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tblib==3.0.0
terminado==0.18.1
tifffile==2023.2.28
tinycss2==1.4.0
tomli==2.0.2
toolz==1.0.0
torch==2.2.2
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xarray==2024.10.0
xmlschema==3.4.3
xsdata==24.3.1
xyzservices==2024.9.0
yarl==1.17.1
zarr==2.15.0
zict==3.0.0
zipp==3.20.2
